+++ src/providers/RoundwareProvider.tsx
@@ -188,8 +188,8 @@ const RoundwareProvider = (props: PropTypes) => {
 		const urlLatitude = searchParams.get('latitude');
 		const urlLongitude = searchParams.get('longitude');
 		const initial_loc = {
-			latitude: parseInt(typeof urlLatitude == 'string' ? urlLatitude : (config.ROUNDWARE_INITIAL_LATITUDE || 0).toString()),
-			longitude: parseInt(typeof urlLongitude == 'string' ? urlLongitude : (config.ROUNDWARE_INITIAL_LONGITUDE || 0).toString()),
+			latitude: parseFloat(typeof urlLatitude == 'string' ? urlLatitude : (config.ROUNDWARE_INITIAL_LATITUDE || 0).toString()),
+			longitude: parseFloat(typeof urlLongitude == 'string' ? urlLongitude : (config.ROUNDWARE_INITIAL_LONGITUDE || 0).toString()),
 		};
 
 		const roundwareOptions: IRoundwareConstructorOptions = {
